회원가입 및 회원정보 수정 페이지 동기화
