+++ src/pages/Signup.tsx
@@ -1,12 +1,15 @@
-import React  from "react";
+import React from "react";
 import { SimpleTemplate } from "./PageTemplate";
-import { SignupContents } from "../components/SignupContents";
+import { UserInfoFrame } from "../components/UserInfoFrame";
+import { UserInfo } from "../types/UserInfo";
+
+
 
 export const Signup: React.FC = () => {
     return (
         <div className="App container">
             <SimpleTemplate param={{ pageHeaderName: "회원가입" }}>
-                <SignupContents />
+                <UserInfoFrame name={""} email={""} userprofilePic={null} isEdit={false} />
             </SimpleTemplate>
         </div>
     );
